requirements.txt
aiogram==3.13.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
apscheduler==3.10.4
pydantic-settings==2.6.1
Pillow>=10.0.0
